generatorAPI/requirements.txt
asn1crypto==1.5.1
bip-utils==2.9.3
blinker==1.8.2
cbor2==5.6.4
cffi==1.17.0
click==8.1.7
coincurve==20.0.0
crcmod==1.7
distlib==0.3.8
ecdsa==0.19.0
ed25519-blake2b==1.4.1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mnemonic==0.21
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.20.0
PyNaCl==1.5.0
pypng==0.20220715.0
qrcode==7.4.2
six==1.16.0
typing_extensions==4.12.2
virtualenv==20.26.3
Werkzeug==3.0.4
